+++ src/components/ChartsContainer.tsx
@@ -40,6 +40,7 @@ const ChartsContainer = ({
         onChange={setUnitPreferences}
         defaultValues={unitPreferences}
       />
+
       <div
         className={cn(
           "tw:w-full tw:relative tw:bg-slate-100 tw:border-slate-700 tw:max-w-[1340px] tw:h-auto tw:mx-auto tw:pr-2 tw:md:px-4 tw:py-0 tw:rounded-lg tw:border tw:overflow-hidden"
